requirement.txt
certifi==2023.5.7
charset-normalizer==2.0.12
cycler==0.11.0
Cython==0.29.35
dataclasses==0.8
decorator==5.1.1
gevent==22.10.2
greenlet==2.0.2
idna==3.4
ipython-genutils==0.2.0
kiwisolver==1.3.1
numpy==1.19.5
onnxruntime-gpu==1.6.0
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.4.0
pip==21.3.1
protobuf==3.19.6
PyAudio==0.2.13
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.27.1
setuptools==58.0.4
six==1.16.0
SpeechRecognition==3.9.0
torch==1.8.0
torchvision==0.9.0
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.16
webrtcvad==2.0.10
websocket==0.2.1
websocket-client==1.3.1
wheel==0.37.1
zope.event==4.6
zope.interface==5.5.2
